More code movement and re-structuring, for better or for worse
