+++ src/CMakeLists.txt
@@ -2,9 +2,10 @@ add_subdirectory(gpu-api)
 add_subdirectory(timing)
 add_subdirectory(matrix_ops)
 add_subdirectory(methods)
+add_subdirectory(planning)
 
 add_library(rtatblas INTERFACE)
 target_include_directories(rtatblas INTERFACE ${CMAKE_CURRENT_LIST_DIR})
-target_link_libraries(rtatblas INTERFACE gpu-api timing matrix_ops methods nlohmann_json::nlohmann_json)
+target_link_libraries(rtatblas INTERFACE gpu-api timing matrix_ops methods planning nlohmann_json::nlohmann_json)
 
 #add_subdirectory(app)
