+++ src/methods/gemm.h
@@ -2,6 +2,7 @@
 #include <string>
 #include <vector>
 #include <matrixop.h>
+#include "executor.h"
 #include "base_options.h"
 
 namespace rtat {
@@ -72,6 +73,34 @@ struct GEMM_Options {
   friend std::istream& operator>>(std::istream&, GEMM_Options&); 
 
   std::unique_ptr<MatrixOp> form_operation(GEMM_Inputs);
+};
+
+
+class GEMM_Executor : public Executor<GEMM_Inputs, GEMM_Key, GEMM_Options> {
+protected:
+  void warmup(GEMM_Inputs params, [[maybe_unused]] GEMM_Options opts,
+              [[maybe_unused]] Stream s) override {
+    size_t n = 8;
+    double *A, *B, *C;
+    gpuAssert(cudaMalloc(&A, n*n*sizeof(double)));
+    gpuAssert(cudaMalloc(&B, n*n*sizeof(double)));
+    gpuAssert(cudaMalloc(&C, n*n*sizeof(double)));
+
+    std::vector<cublasOperation_t> ops = {CUBLAS_OP_N, CUBLAS_OP_T};
+
+    for (auto &opA : ops) {
+      for (auto &opB : ops) {
+        double alpha = 1.0;
+        double beta = 0.0;
+        cublasDgemm(params.handle, opA, opB, n,n,n, &alpha, A,n,B,n, &beta, C,n);
+        cublasDgeam(params.handle, opA, opB, n,n, &alpha, A, n, &beta, B, n, C, n);
+      }
+    }
+    gpuAssert(cudaDeviceSynchronize());
+    gpuAssert(cudaFree(A));
+    gpuAssert(cudaFree(B));
+    gpuAssert(cudaFree(C));
+  }
 };
 
 }
