+++ src/methods/predicates.h
